+++ openapi_server/models/address.py
@@ -16,7 +16,7 @@ class Address(Model):
     Do not edit the class manually.
     """
 
-    def __init__(self, address: str=None, balance: Values=None, currency: str=None, entity: int=None, first_tx: TxSummary=None, in_degree: int=None, last_tx: TxSummary=None, no_incoming_txs: int=None, no_outgoing_txs: int=None, out_degree: int=None, total_received: Values=None, total_spent: Values=None):
+    def __init__(self, address: str=None, balance: Values=None, currency: str=None, entity: int=None, first_tx: TxSummary=None, in_degree: int=None, last_tx: TxSummary=None, no_incoming_txs: int=None, no_outgoing_txs: int=None, out_degree: int=None, status: str=None, total_received: Values=None, total_spent: Values=None):
         """Address - a model defined in OpenAPI
 
         :param address: The address of this Address.
@@ -29,6 +29,7 @@ class Address(Model):
         :param no_incoming_txs: The no_incoming_txs of this Address.
         :param no_outgoing_txs: The no_outgoing_txs of this Address.
         :param out_degree: The out_degree of this Address.
+        :param status: The status of this Address.
         :param total_received: The total_received of this Address.
         :param total_spent: The total_spent of this Address.
         """
@@ -43,6 +44,7 @@ class Address(Model):
             'no_incoming_txs': int,
             'no_outgoing_txs': int,
             'out_degree': int,
+            'status': str,
             'total_received': Values,
             'total_spent': Values
         }
@@ -58,6 +60,7 @@ class Address(Model):
             'no_incoming_txs': 'no_incoming_txs',
             'no_outgoing_txs': 'no_outgoing_txs',
             'out_degree': 'out_degree',
+            'status': 'status',
             'total_received': 'total_received',
             'total_spent': 'total_spent'
         }
@@ -72,6 +75,7 @@ class Address(Model):
         self._no_incoming_txs = no_incoming_txs
         self._no_outgoing_txs = no_outgoing_txs
         self._out_degree = out_degree
+        self._status = status
         self._total_received = total_received
         self._total_spent = total_spent
 
@@ -102,6 +106,7 @@ class Address(Model):
             'no_incoming_txs': self._no_incoming_txs,
             'no_outgoing_txs': self._no_outgoing_txs,
             'out_degree': self._out_degree,
+            'status': self._status,
             'total_received': self._total_received,
             'total_spent': self._total_spent }
 
@@ -341,6 +346,33 @@ class Address(Model):
             raise ValueError("Invalid value for `out_degree`, must not be `None`")
 
         self._out_degree = out_degree
+
+    @property
+    def status(self):
+        """Gets the status of this Address.
+
+
+        :return: The status of this Address.
+        :rtype: str
+        """
+        return self._status
+
+    @status.setter
+    def status(self, status):
+        """Sets the status of this Address.
+
+
+        :param status: The status of this Address.
+        :type status: str
+        """
+        allowed_values = ["clean", "dirty", "new"]  # noqa: E501
+        if status not in allowed_values:
+            raise ValueError(
+                "Invalid value for `status` ({0}), must be one of {1}"
+                .format(status, allowed_values)
+            )
+
+        self._status = status
 
     @property
     def total_received(self):